Moved RandomColor to utils directory
